fix: drag cursor math and unify cross-track routing
