+++ src/views/timeline/drag-end-resolution.ts
@@ -2,7 +2,6 @@ import type { LyricLine } from "@/domain/line/model";
 import { GROUP_HEADER_HEIGHT } from "@/views/timeline/group-header-row";
 import { GUTTER_WIDTH, useTimelineStore, WAVEFORM_HEIGHT } from "@/views/timeline/timeline-store";
 import { computeRowLayout, getLineAndTrackAtY } from "@/views/timeline/utils";
-import type { DragEndEvent } from "@dnd-kit/core";
 
 // -- Constants -----------------------------------------------------------------
 
@@ -18,20 +17,27 @@ interface DropTarget {
   cursorTime: number;
 }
 
+interface ResolveDropTargetInput {
+  clientX: number;
+  clientY: number;
+  lines: LyricLine[];
+}
+
 // -- Helpers -------------------------------------------------------------------
 
-function resolveDropTarget(event: DragEndEvent, lines: LyricLine[]): DropTarget | null {
-  const { activatorEvent, delta } = event;
-  if (!(activatorEvent instanceof PointerEvent)) return null;
-
+// clientX/Y are viewport-relative live pointer coordinates (from a pointermove
+// listener), not derived from dnd-kit's delta. dnd-kit's delta is already
+// scroll-adjusted, so combining it with scrollTop double-counts auto-scroll.
+// Live pointer coordinates already reflect any scroll-during-drag, so adding
+// container.scrollTop here cleanly converts viewport to container-content
+// coordinates without double-counting.
+function resolveDropTarget({ clientX, clientY, lines }: ResolveDropTargetInput): DropTarget | null {
   const container = document.querySelector<HTMLDivElement>("[data-scroll-container]");
   if (!container) return null;
 
   const rect = container.getBoundingClientRect();
-  const cursorClientX = activatorEvent.clientX + delta.x;
-  const cursorClientY = activatorEvent.clientY + delta.y;
-  const cursorX = cursorClientX - rect.left + container.scrollLeft;
-  const cursorY = cursorClientY - rect.top + container.scrollTop;
+  const cursorX = clientX - rect.left + container.scrollLeft;
+  const cursorY = clientY - rect.top + container.scrollTop;
 
   const { zoom, rowHeights, defaultRowHeight, collapsedInstances } = useTimelineStore.getState();
   const layout = computeRowLayout({
