+++ src/views/timeline/utils.hit-test.test.ts
@@ -52,4 +52,42 @@ describe("getLineAndTrackAtY", () => {
       }
     });
   });
+
+  describe("hit test boundaries", () => {
+    const firstPos = layout.lineTops.get("l1");
+    const secondPos = layout.lineTops.get("l2");
+    if (!firstPos || !secondPos) throw new Error("missing layout positions");
+
+    it("returns first row at y equal to top of first row", () => {
+      expect(getLineAndTrackAtY(firstPos.top, lines, layout)).toEqual({ lineIndex: 0, track: "word" });
+    });
+
+    it("returns first row at the last pixel of the first row (top + height - 1)", () => {
+      const y = firstPos.top + firstPos.height - 1;
+      const hit = getLineAndTrackAtY(y, lines, layout);
+      expect(hit?.lineIndex).toBe(0);
+    });
+
+    it("returns next row at y one past the bottom of the first row (top + height)", () => {
+      const y = firstPos.top + firstPos.height;
+      const hit = getLineAndTrackAtY(y, lines, layout);
+      expect(hit).toEqual({ lineIndex: 1, track: "word" });
+      expect(y).toBe(secondPos.top);
+    });
+
+    it("returns null at one past the last pixel of the last row", () => {
+      const y = secondPos.top + secondPos.height;
+      expect(getLineAndTrackAtY(y, lines, layout)).toBeNull();
+    });
+
+    it("returns word track at the last pixel of the main half (mainBottom - 1)", () => {
+      const y = firstPos.mainBottom - 1;
+      expect(getLineAndTrackAtY(y, lines, layout)).toEqual({ lineIndex: 0, track: "word" });
+    });
+
+    it("returns bg track exactly at mainBottom (the seam belongs to bg)", () => {
+      const y = firstPos.mainBottom;
+      expect(getLineAndTrackAtY(y, lines, layout)).toEqual({ lineIndex: 0, track: "bg" });
+    });
+  });
 });
